Fix bug in decode_critical_parameter()
The function iterates on the critical parameters in a CBOR array,
but after checking the QCBORDecode_GetNext return value, always exits
the while loop without adding any elements to the list.

This error would have been caught by the test case crit_parameters_test,
but this one also had a bug. If run_test_sign_and_verify returned
with an unexpected value for the first item in the crit_tests_table,
crit_parameters_test returned 0. This was interpreted as test pass by
the caller.

Change-Id: Idc6e699c6b794704f35b5e4c52eed5a3adc9e0a6
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>