Add crypto adapter context to sign context
This commit defines a new structure called t_cose_crypto_backend_ctx.
The t_cose_sign1_sign_ctx structure now holds a pointer to this new
context.

The idea behind this change is that the crypto adapation layer might
need resources that are provided by the caller of the t_cose library.

The t_cose_crypto_backend_ctx structure belongs to the crypto adaption
layer, but should be available in the source calling the t_cose library,
so a new public header is added containing this structure.

Change-Id: I4f910c969fb08142b2e24e0dc2acf383e3d253a4
Co-authored-by: Tamas Ban <tamas.ban@arm.com>
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>