Move hash context to the crypto context
This commit moves the fields from `struct t_cose_crypto_hash` to the
crypto context structure of the related crypto adappter. This makes
possible to compile t_cose with a crypto adapter that is defined outside
of the t_cose source tree.

This makes allocating a crypto context mandatory for all the crpto
adapters that are defined in t_cose, so instead of requiring calling a
separate setter for the context, a new `crypto_context` parameters is
added to the `t_cose_sign1_sign_init` and `t_cose_sign1_verify_init`
structures.

If a crypto adapter doesn't require context to be stored, the caller
still can pass NULL in the init functions.

Change-Id: Iba84febf688a6838b362feaf51ee42adfff021f2