Add an option to skip checking the critical header parameter.
If the enveloppe contains critical header parameters beyond those
understood by t_cose, verification would fail. This precludes the use of
application-specific critical header parameters.

Adding a flag to skip this step of the validation allows applications to
support such parameters. It is up to the application to replicate the
critical header parameters verification feature, as necessary.

Co-authored-by: Joe Powell <joepowell@microsoft.com>