Introduce T_COSE_DISABLE_COSE_SIGN option
Introduce the T_COSE_DISABLE_COSE_SIGN option and replace
T_COSE_DISABLE_SIGN1 with it. The tests that use multiple
signatures can also be disabled with this option.
COSE_Sign1 structure support is always enabled.

The purpose of this change is to achieve an effective code size
reduction by disabling a subset of features - when needed and
other linker optimization features are less efficient.

Change-Id: I83a33c34d4e14c86274ce0cfd1115e8b0961a7a3
Signed-off-by: David Vincze <david.vincze@arm.com>